@@ -4,9 +4,10 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
+  bloc: ^8.1.0
   bloc_concurrency: ^0.2.0
   equatable: ^2.0.3
   flutter:
@@ -19,9 +20,7 @@ dev_dependencies:
   bloc_test: ^9.0.0
   flutter_test:
     sdk: flutter
-  mocktail: ^0.3.0
-  very_good_analysis: ^3.0.1
-  flutter_lints: ^2.0.0
+  mocktail: ^1.0.0
 
 flutter:
   uses-material-design: true
